feat(login): портировать логин-сервер под протокол клиента 3.5.0.3
Перенесены изменения AAEmu.Login из проекта 3.5.0.3 (AAC3500) по материалам
побайтового сравнения 1.2 -> 3.5 (diff-3503); патчи применены выборочно
с --include='AAEmu.Login/**':

- LoginClient: новые сигнатуры ACJoinResponsePacket (слоты персонажей 6->2,
  AdditionalData=22) и ACAuthResponsePacket под клиента 3.5;
- LoginSession: переработанная state-машина логина;
- CARequestAuthTencentPacketHandler: рабочая аутентификация через
  KoreaAuthFlowFactory вместо no-op заглушки;
- добавлены CARequestAuth_004_Packet и его обработчик;
- структуры пакетов C2L/L2C и опкоды (CLOffsets, LCOffsets, GLOffsets,
  LGOffsets) синхронизированы с клиентом 3.5.0.3;
- добавлены CHANGELOG.MD и Docs/LoginServer3503Port_ru.md с описанием порта.

Сознательно не перенесено (следующие шаги по PORTING-1.2-to-3.5.md):
AccountId uint->ulong в AAEmu.Commons (затрагивает Game-сервер), game-часть
протокола 3.5 (шифрование, GamePacket level 5, X2EnterWorld*).

Проверено: сборка AAEmu.Login, AAEmu.UnitTests и AAEmu.Login.IntegrationTests
без ошибок; 1076/1076 unit-тестов успешно.

+++ AAEmu.Login/Core/Packets/C2L/CARequestReconnectPacket.cs
@@ -30,8 +30,8 @@ public class CARequestReconnectPacket() : LoginPacket(TypeId), ILoginPacket
     {
         var pFrom = stream.ReadUInt32();
         var pTo = stream.ReadUInt32();
-        AccountId = new AccountId(stream.ReadUInt32());
-        GsId = new GameServerId(stream.ReadByte());
+        AccountId = new AccountId(stream.ReadUInt64());
+        GsId = new GameServerId(stream.ReadByte()); // wid
         Cookie = stream.ReadUInt32();
         var macLength = stream.ReadUInt16();
         var mac = stream.ReadBytes(macLength);
